requirements.txt
pandas
numpy
requests
yfinance

streamlit
plotly
pyyaml

scikit-learn
backtrader
pyfolio